WCM-285: move Converter back to filesystem
