+++ core/src/zeit/connector/filesystem.py
@@ -6,8 +6,11 @@ import logging
 import os
 import os.path
 
+from sqlalchemy.dialects.postgresql import JSONB
 import gocept.cache.property
+import grokcore.component as grok
 import lxml.etree
+import sqlalchemy
 import zope.app.file.image
 import zope.interface
 
@@ -268,7 +271,7 @@ class Connector:
             column = Content.column_by_name(*key)
             if column is None:
                 continue
-            converter = zeit.connector.interfaces.IConverter(column)
+            converter = IConverter(column)
             properties[key] = converter.deserialize(value)
 
     def _guess_type(self, id):
@@ -316,3 +319,75 @@ def parse_properties(xml):
         value += '</tag:rankedTags>'
         properties[('keywords', 'http://namespaces.zeit.de/CMS/tagging')] = value
     return properties
+
+
+class IConverter(zope.interface.Interface):
+    """Converts webdav values to and from the postgresql database."""
+
+    def serialize(value):
+        pass
+
+    def deserialize(value):
+        pass
+
+
+@grok.implementer(IConverter)
+@grok.adapter(sqlalchemy.Column)
+def converter_from_column_type(column):
+    return IConverter(column.type)
+
+
+@grok.implementer(IConverter)
+class DefaultConverter(grok.Adapter):
+    grok.context(zope.interface.Interface)
+
+    def serialize(self, value):
+        return value
+
+    def deserialize(self, value):
+        return value
+
+
+class BoolConverter(DefaultConverter):
+    grok.context(sqlalchemy.Boolean)
+
+    def serialize(self, value):
+        return zeit.cms.content.dav.BoolProperty._toProperty(value)
+
+    def deserialize(self, value):
+        return zeit.cms.content.dav.BoolProperty._fromProperty(value)
+
+
+class IntConverter(DefaultConverter):
+    grok.context(sqlalchemy.Integer)
+
+    def serialize(self, value):
+        return str(value)
+
+    def deserialize(self, value):
+        return int(value)
+
+
+class DatetimeConverter(DefaultConverter):
+    grok.context(sqlalchemy.TIMESTAMP)
+
+    def serialize(self, value):
+        return zeit.cms.content.dav.DatetimeProperty._toProperty(value)
+
+    def deserialize(self, value):
+        return zeit.cms.content.dav.DatetimeProperty._fromProperty(value)
+
+
+@grok.implementer(IConverter)
+class DictConverter(grok.Adapter):
+    grok.context(JSONB)
+
+    def serialize(self, value):
+        if not value:
+            return {}
+        return value
+
+    def deserialize(self, value):
+        if not value:
+            return {}
+        return value
